refactor(supervisor): extract mechanism code into leaf modules
Pull the low-level mechanism out of the supervisor so it reads as
orchestration only:

- native_input.rs: NativeInputTracker owns the native-view line
  discipline (backspace/ctrl-C/escape-sequence/UTF-8 parsing) that was
  inlined in the supervisor; now unit-tested for the first time.
- transcripts.rs: transcript locating/reading (claude projects glob,
  codex rollouts walk, JSONL reader) moves out, fixing agent.rs's
  upward dependency on supervisor for pure fs helpers.
- WorkspaceManager::project_setting(): the Run panel's project-settings
  lookup replaces inline rusqlite SQL in read_run_commands; the now
  unused db_handle() escape hatch is removed.
- choose_restore_branch_name(): the -restored suffix collision loop is
  lifted out of restore_agent.

No behavior change; tests move with their code.
