+++ src-tauri/src/transcripts.rs
@@ -1,0 +1,214 @@
+//! Locating and reading provider transcript files on disk.
+//!
+//! Pure filesystem mechanism — no supervisor or workspace state. The
+//! per-provider reader tables in `agent.rs` point their `locate`/`read`
+//! hooks here.
+
+use std::path::{Path, PathBuf};
+
+use serde_json::Value;
+
+use crate::error::{Error, Result};
+
+/// Locate the claude session JSONL by scanning the candidate `projects/*/`
+/// dirs (see [`claude_projects_dirs`]) for `<session-id>.jsonl`. Claude's
+/// path-encoding scheme isn't part of its public API, so we glob instead of
+/// recomputing the encoded directory name from the worktree path.
+pub(crate) fn find_session_jsonl(session_id: &str) -> Option<PathBuf> {
+    find_session_jsonl_in(&claude_projects_dirs(), session_id)
+}
+
+/// Candidate `projects` directories Claude may have written transcripts to.
+/// Honors `CLAUDE_CONFIG_DIR` (Claude CLI's own config-dir override) and always
+/// also includes the default `~/.claude`, so a transcript is located regardless
+/// of which config dir was active when the agent was spawned. Mirrors the way
+/// `find_codex_rollouts` honors `CODEX_HOME` — without this, an agent spawned
+/// with `CLAUDE_CONFIG_DIR` set wrote its transcript somewhere we never scanned,
+/// so it was never ingested into `session_records` and was lost when that dir
+/// moved.
+fn claude_projects_dirs() -> Vec<PathBuf> {
+    projects_dirs_from(
+        std::env::var_os("CLAUDE_CONFIG_DIR").map(PathBuf::from),
+        dirs::home_dir(),
+    )
+}
+
+/// Pure core of [`claude_projects_dirs`]: configured dir first (if any), then
+/// the default `~/.claude`, deduped so an explicit `CLAUDE_CONFIG_DIR=~/.claude`
+/// doesn't double-scan.
+fn projects_dirs_from(config_dir: Option<PathBuf>, home: Option<PathBuf>) -> Vec<PathBuf> {
+    let mut out: Vec<PathBuf> = Vec::new();
+    let mut push = |base: PathBuf| {
+        let p = base.join("projects");
+        if !out.contains(&p) {
+            out.push(p);
+        }
+    };
+    if let Some(cfg) = config_dir {
+        push(cfg);
+    }
+    if let Some(home) = home {
+        push(home.join(".claude"));
+    }
+    out
+}
+
+/// Scan the given `projects` dirs for `<session-id>.jsonl`, returning the first
+/// match. A missing/unreadable candidate dir is skipped, not fatal, so a later
+/// candidate is still searched.
+fn find_session_jsonl_in(projects_dirs: &[PathBuf], session_id: &str) -> Option<PathBuf> {
+    let filename = format!("{session_id}.jsonl");
+    for projects in projects_dirs {
+        let Ok(entries) = std::fs::read_dir(projects) else {
+            continue;
+        };
+        for entry in entries.flatten() {
+            let path = entry.path().join(&filename);
+            if path.exists() {
+                return Some(path);
+            }
+        }
+    }
+    None
+}
+
+/// All of codex's rollout files for a thread id, ordered (filenames are
+/// timestamp-prefixed, so lexical sort == chronological). Codex stores sessions
+/// at `$CODEX_HOME/sessions/YYYY/MM/DD/rollout-<ts>-<id>.jsonl` (CODEX_HOME
+/// defaults to `~/.codex`); the id suffix is the thread id we captured. Resume
+/// normally keeps one file per session, but returning all is correct if it splits.
+pub(crate) fn find_codex_rollouts(session_id: &str) -> Vec<PathBuf> {
+    let Some(home) = std::env::var_os("CODEX_HOME")
+        .map(PathBuf::from)
+        .or_else(|| dirs::home_dir().map(|h| h.join(".codex")))
+    else {
+        return Vec::new();
+    };
+    // Anchor on the `-<id>.jsonl` boundary (filenames are
+    // `rollout-<ts>-<id>.jsonl`) so one thread id can't match another whose
+    // name merely ends with the same characters.
+    let suffix = format!("-{session_id}.jsonl");
+    // Walk the YYYY/MM/DD tree (three dir levels) and match the suffix.
+    fn dirs_in(p: &Path) -> Vec<PathBuf> {
+        std::fs::read_dir(p)
+            .into_iter()
+            .flatten()
+            .flatten()
+            .map(|e| e.path())
+            .filter(|p| p.is_dir())
+            .collect()
+    }
+    let sessions = home.join("sessions");
+    let mut out = Vec::new();
+    for year in dirs_in(&sessions) {
+        for month in dirs_in(&year) {
+            for day in dirs_in(&month) {
+                for entry in std::fs::read_dir(&day).into_iter().flatten().flatten() {
+                    let path = entry.path();
+                    if path
+                        .file_name()
+                        .and_then(|n| n.to_str())
+                        .is_some_and(|n| n.ends_with(&suffix))
+                    {
+                        out.push(path);
+                    }
+                }
+            }
+        }
+    }
+    out.sort();
+    out
+}
+
+/// Read a JSONL file into a vec of parsed values, skipping blank or
+/// unparseable lines.
+pub(crate) fn read_jsonl_values(path: &Path) -> Result<Vec<Value>> {
+    use std::io::BufRead;
+    let file = std::fs::File::open(path)
+        .map_err(|e| Error::Other(format!("open transcript: {e}")))?;
+    let reader = std::io::BufReader::new(file);
+    let mut out = Vec::new();
+    for line in reader.lines().map_while(std::result::Result::ok) {
+        if line.trim().is_empty() {
+            continue;
+        }
+        if let Ok(v) = serde_json::from_str::<Value>(&line) {
+            out.push(v);
+        }
+    }
+    Ok(out)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    // ── claude transcript location (CLAUDE_CONFIG_DIR) ────────────────────────
+
+    #[test]
+    fn projects_dirs_prefers_config_dir_then_default_home() {
+        let dirs = projects_dirs_from(
+            Some(PathBuf::from("/home/u/.claude-eve")),
+            Some(PathBuf::from("/home/u")),
+        );
+        assert_eq!(
+            dirs,
+            vec![
+                PathBuf::from("/home/u/.claude-eve/projects"),
+                PathBuf::from("/home/u/.claude/projects"),
+            ],
+        );
+    }
+
+    #[test]
+    fn projects_dirs_dedups_when_config_is_default() {
+        // CLAUDE_CONFIG_DIR explicitly set to ~/.claude must not double-scan.
+        let dirs = projects_dirs_from(
+            Some(PathBuf::from("/home/u/.claude")),
+            Some(PathBuf::from("/home/u")),
+        );
+        assert_eq!(dirs, vec![PathBuf::from("/home/u/.claude/projects")]);
+    }
+
+    #[test]
+    fn projects_dirs_falls_back_to_home_when_config_unset() {
+        let dirs = projects_dirs_from(None, Some(PathBuf::from("/home/u")));
+        assert_eq!(dirs, vec![PathBuf::from("/home/u/.claude/projects")]);
+    }
+
+    #[test]
+    fn find_session_jsonl_locates_transcript_in_relocated_config_dir() {
+        // Regression: the locator used to hardcode `~/.claude/projects`, so an
+        // agent spawned with CLAUDE_CONFIG_DIR pointing elsewhere wrote its
+        // transcript to a dir we never scanned — it was never ingested and was
+        // lost when that dir moved. The transcript must be found wherever the
+        // configured projects dir is.
+        let cfg = tempfile::tempdir().unwrap();
+        let projects = cfg.path().join("projects");
+        let slug = projects.join("-Users-alex--fletch-worktrees-transylvania-fletch");
+        std::fs::create_dir_all(&slug).unwrap();
+        let sid = "f90f9c57-6dd1-45a0-9b69-5b5963979d5b";
+        let jsonl = slug.join(format!("{sid}.jsonl"));
+        std::fs::write(&jsonl, b"{}\n").unwrap();
+
+        let found = find_session_jsonl_in(&[projects], sid);
+        assert_eq!(found.as_deref(), Some(jsonl.as_path()));
+    }
+
+    #[test]
+    fn find_session_jsonl_skips_missing_dir_and_scans_the_next() {
+        // A non-existent candidate dir (e.g. the default ~/.claude when only the
+        // relocated config dir has the file) must not short-circuit the scan.
+        let cfg = tempfile::tempdir().unwrap();
+        let projects = cfg.path().join("projects");
+        let slug = projects.join("slug");
+        std::fs::create_dir_all(&slug).unwrap();
+        let sid = "abc";
+        let jsonl = slug.join(format!("{sid}.jsonl"));
+        std::fs::write(&jsonl, b"{}\n").unwrap();
+
+        let missing = cfg.path().join("does-not-exist");
+        let found = find_session_jsonl_in(&[missing, projects], sid);
+        assert_eq!(found.as_deref(), Some(jsonl.as_path()));
+    }
+}
